refactor: move roundtrip latency test to examples
