--- showcase/latency_tester/Cargo.toml
@@ -1,12 +1,0 @@
-[package]
-name = "latency_tester"
-version = "0.1.0"
-edition = "2021"
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
-web-audio-api = { path = "../../" }
-
-[features]
-jack = ["web-audio-api/cpal-jack"]
